refactor(engine): move class to tests
The class is only used in one test
